refactor: date 체크 로직 분리
related to: #75
